Initial Version
I am not entirely confident that my python structures are correct as some of the structure conventions in the file format seemed obscure. Notably the way strings are defined, which seem to use a special terminator character other then 0. The terminator appears to be an enumerator which results in additional data being defined after the string. This can provide additional attributes that affect other dynamic structures, therefore with the limited samples I received I can only imagine there is more to the structure. That being said I did not complete the mesh import function due to the uncertainty of the file format structure, only the basic geometry is imported at this time.